If there is no valid baseline output to compare against (valid meaning it exists, in that there exists an exit status file, and the exit status was no 139 for the original program), then ps validate will try no input for the transformed program and compare the exit status with the original program.
If the original program segfaults on no input, and no valid baseline output exists, then the transform is validated blindly.

Since validation is done blindly in the above scenario, do_p1transform.sh was changed to check for valid baseline output prior to issuing p1. If no valid baseline output exists, the coverage file is deleted, and replaced with an empty coverage file. This ensures that p1 is performed only, since the abscence of coverage data automatically results in p1 (as opposed to pn).




Former-commit-id: 43822fa92319636e73c0fedf77a40fc2c14bd706