Previous version of ps_validate called replayer with no input in grace failed. This was believed to have worked in the past, but is now broken. It is now called with empty.json saved in peasoup_home.
Also, in case of complete do_concolic script failure, if no inputs are found, a check is made to see if a.sym was produced, if not ps_validate.sh will produce it. 

With no concolic inputs, the previous version did not exit with an error status if the empty input caused a divergence in exit status. 



Former-commit-id: a897ae79f75704319c17638cedc2117ff68b7064