Typo found in PNRegularExpressions for ebp scaled regex, removed a useless + in the regex.
If a scaled memory access is found that uses ebp as an index
e.g., [ecx+ebp*1+0x21]
The funciton is considered not PN safe, and only p1 is attempted if p1
has been determined to be safe. 

This check is done in OffsetInference.cpp. In PNTransformDriver.cpp, if this
pattern is encountered during transformation, it is assumed that p1 is being 
used and the offset is used to find the appropriate displacment by which
to modify the original instruction. 

In the future a dynamic solution would be more appropriate, and then PN
could be attempted. The issue with PN is the offset discovered in instructions
like these does not necessarily point to the variable the instruction will
reference at runtime. A runtime check is needed to see what memory is accessed. 




Former-commit-id: c5a33834377949a5cfef357004bdd29c6bc079c9