The primary change in this commit is the support of dynamic arrays and alloca functions as well as adding a few more prologue checks.
Added General_Utility.*, as of this commit, this utility "class" contains a str2int and str2uint function. 

Changed makefile to reflect addition of General_Utility.*.

Modified PNRegularExpressions.cpp, changed stack_dealloc_implicit. Added the pattern

lea esp, [ebp-<const>]

as a valid stack deallocation. 

OffsetInference.cpp now checks the amount by which esp is subtracted in 

sub esp, <val>

instructions. If the value is determined to be non-integral, it is assumed the function is using alloca or dynamic arrays, and the function is considered non PNsafe (only p1 is attempted). 

Given the addition to PNRegularExpressions.cpp, in OffsetInference.cpp, stack_dealloc_implicit checks are now done first. 

PNTransformDriver.cpp also does a check to see if sub esp, <val> contains an integral value, and does not transform instructions that do not. 

OffsetInference.cpp now does a stack deallocation check for add esp, 0x00000000, and does not filter a function based on 0 not matching the size of the stack frame. 




Former-commit-id: 9c9f3494d7bd13bf237ef9fc4925b1a72e95b3c2