PN optimization. PN will now do a binary search algorithm to find bad modifications. All functions with coverage are now validated, and validated faster than previous version that only validated if coverage was above the threshold.
This version does not have shuffle validation, I will add that feature in a future update. It was easier to remove it for this optimization. Typically this is only used for Main. 

New logging summary output including how many functions are modified with high coverage, low coverage and no coverage. 



Former-commit-id: 468864bd92166e2eafafddd65b152c721173d294