Update Zipr to support plugins plopping insns.
1. Add support for plugins to call certain
Zipr_t functions. This meant making a class
that is Zipr_t and then converting the actual
implementation to a subclass (ZiprImpl_t). The
Zipr_t goes in the zipr_sdk.

2. Ask all plugins about each instruction
and determine whether there are any takers
for plopping.

3. When plopping an instruction, consider
whether or not a plugin asked to do the plopping.
If so, let it do the work.