Bugfix: Small fallthrough dollops that are placed cause coalesce problems
When a dollop with small fallthroughs would fit entirely in an 
allocated region but some of those fallthroughs are already placed,
coalescing would miscalculate what to do. It would say, "Let's
place it all" but would eventually want to put a jmp to the 
already-placed dollop. When the fallthoughs were small, that 
jump would be bigger than the entirety of dollop and cause
us to overflow the reserved space.