Fix #5: (Related) Could not find pin for EH Table
This fixes an issue where a pinned address was

1. unpinned
2. reserved as part of a chain
3. exists in a sled

and all hell broke loose. The fix is to make sure
that the pinned addresses are only recorded *after*
the initial unpinning code happens.

As a further fix, there should be a function that
unpins an address (it would reset the ibta of
the instruction and clear the pinned address map.)