Rework the assembly-string->binary conversion.
The old method involved writing assembly to a file, asking an external assembler to do the conversion, then processing the binary file to read the binary instructions back out.  This method is slow and relies on nasm, which isn't available on all machines or for the full variety of supported architectures.  The replacement is keystone, a 3rd party library that is linked to provide assembly->binary conversion.