Fix buggy unpinner when unpin stops early
Debugging code allows for one to binary-search the unpins.

But a .dynsym, .got, etc. entry can be duplicated in it's own section
or other sections.  Unpinning has to handle unpinning an instruction
and deal with updating all tables properly.  Bug was found in handling
of the unpin for .dynsym in this regard.  This patch fixes that bug.