Zipr/FII failure-to-tranform fixes
1) FII was finding the wrong cmp for switch tables.
Removed the blind cmp and forced the cmp to use the proper
register or not find a cmp.  Also changed the limit for a table
size when no cmp is found from 2^32 to 255.

2) Fixed an issue with zipr dollop layout where the size
to layout 1 instruction + a trampoline was bigger than laying
out two+ full dollops (because both dollops were short), but
the dollops are disallowed from being coalesced.  The worst-size
detector needed to account for non-coalescable dollops.