Update to fix calls for exception handling.
Fixed calls break exception handling, and largely aren't needed anymore.
We were still "fixing" calls when it had no fallthrough instruction,
assuming the call is used for something else.  Except that sometimes a
"call _cxa_throw" ends a function/section and legitimately has no
fallthrough.  And this breaks the exception handling.

Testing this fix against the Zipr test suite to verify that it is OK
on a large suite of programs.